kloud: move Credential models to koding/models pkg
